[Heap*.h/cpp] Refactor: Update whitespace and indentation;  rename files
