+++ src/database/heap_file_encoder.cpp
@@ -1,0 +1,14 @@
+#include "heap_file_encoder.h"
+
+namespace emerald {
+HeapFileEncoder::HeapFileEncoder();
+
+static void HeapFileEncoder::convert(std::vector << std::vector<int>> tuples, FILE *outFile, int npagebytes, int numFields) {}
+
+static void HeapFileEncoder::convert(FILE *inFile, FILE *outFile, int npagebytes, int numFields) {}
+
+static void HeapFileEncoder::convert(FILE *inFile, FILE *outFile, int npagebytes, int numFields, types[] typeAr) {}
+
+static void HeapFileEncoder::convert(FILE *inFile, FILE *outFile, int npagebytes, int numFields, types[] typeAr, char fieldSeparator) {}
+
+}
